type fixes and make class open

+++ sdk/src/main/java/sdk/api/core_api_provider.kt
@@ -47,7 +47,7 @@ class CoreApiProvider(
         ) as BasicResponse<List<Map<String, Any>>>
     }
 
-    suspend fun getAllStocks(region: Region, secondsSinceEpoch: Int? = null): BasicResponse<List<Map<String, Any>>> {
+    suspend fun getAllStocks(region: Region, secondsSinceEpoch: Long? = null): BasicResponse<List<Map<String, Any>>> {
         val path = "stock/all/${region.string()}"
         val response = httpHandler.get(
             path = path,
